식재료 교환 엔티티, 회원 엔티티 관계 연결 추가

+++ src/main/java/refooding/api/domain/exchange/service/ExchangeServiceImpl.java
@@ -5,6 +5,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import refooding.api.common.exception.CustomException;
+import refooding.api.common.exception.ExceptionCode;
 import refooding.api.domain.exchange.dto.request.ExchangeCreateRequest;
 import refooding.api.domain.exchange.dto.request.ExchangeUpdateRequest;
 import refooding.api.domain.exchange.dto.response.ExchangeDetailResponse;
@@ -15,6 +17,8 @@ import refooding.api.domain.exchange.entity.Region;
 import refooding.api.domain.exchange.repository.ExchangeRepository;
 import refooding.api.domain.exchange.repository.ExchangeSearchCondition;
 import refooding.api.domain.exchange.repository.RegionRepository;
+import refooding.api.domain.member.entity.Member;
+import refooding.api.domain.member.repository.MemberRepository;
 
 @Service
 @Transactional(readOnly = true)
@@ -23,9 +27,10 @@ public class ExchangeServiceImpl implements ExchangeService{
 
     private final ExchangeRepository exchangeRepository;
     private final RegionRepository regionRepository;
+    private final MemberRepository memberRepository;
 
     @Override
-    public Slice<ExchangeResponse> findExchanges(String keyword, ExchangeStatus status, Long regionId, Long lastExchangeId, Pageable pageable) {
+    public Slice<ExchangeResponse> getExchanges(String keyword, ExchangeStatus status, Long regionId, Long lastExchangeId, Pageable pageable) {
         return exchangeRepository.findExchangeByCondition(
                 new ExchangeSearchCondition(keyword, status, regionId, lastExchangeId),
                 pageable
@@ -33,22 +38,41 @@ public class ExchangeServiceImpl implements ExchangeService{
     }
 
     @Override
-    @Transactional
-    public Long create(ExchangeCreateRequest request) {
-        // TODO : 에러처리
-        Region region = regionRepository.findById(request.regionId()).orElseThrow();
-        Exchange exchange = request.toExchange(region);
-
-        exchangeRepository.save(exchange);
-        return exchange.getId();
+    public ExchangeDetailResponse getExchangeById(Long exchangeId) {
+        Exchange exchange = getById(exchangeId);
+        return ExchangeDetailResponse.from(exchange);
     }
 
     @Override
     @Transactional
+    public Long create(ExchangeCreateRequest request) {
+        // TODO : 회원 도메인 구현시 적용
+        // 임시 회원 아이디
+        Long memberId = 1L;
+        Member findMember = getMemberById(memberId);
+        Region region = getRegionById(request.regionId());
+
+        Exchange exchange = request.toExchange(region, findMember);
+        exchangeRepository.save(exchange);
+
+        return exchange.getId();
+    }
+
+
+
+    @Override
+    @Transactional
     public void update(Long exchangeId, ExchangeUpdateRequest request) {
-        // TODO : 에러처리
-        Exchange exchange = exchangeRepository.findById(exchangeId).orElseThrow();
-        Region region = regionRepository.findById(request.regionId()).orElseThrow();
+        // TODO : 회원 도메인 구현시 적용
+        // 임시 회원 아이디
+        Long memberId = 1L;
+        Member findMember = getMemberById(memberId);
+
+        Exchange exchange = getById(exchangeId);
+        if (exchange.validateMember(findMember.getId())) {
+            throw new CustomException(ExceptionCode.UNAUTHORIZED);
+        }
+        Region region = getRegionById(request.regionId());
 
         exchange.updateExchange(
                 request.title(),
@@ -61,16 +85,31 @@ public class ExchangeServiceImpl implements ExchangeService{
     @Override
     @Transactional
     public void delete(Long exchangeId) {
-        // TODO : 에러처리
-        Exchange exchange = exchangeRepository.findById(exchangeId).orElseThrow();
+        // TODO : 회원 도메인 구현시 적용
+        // 임시 회원 아이디
+        Long memberId = 1L;
+        Member findMember = getMemberById(memberId);
+
+        Exchange exchange = getById(exchangeId);
+        if (exchange.validateMember(findMember.getId())) {
+            throw new CustomException(ExceptionCode.UNAUTHORIZED);
+        }
+
         exchange.delete();
     }
 
-    @Override
-    public ExchangeDetailResponse getExchangeById(Long exchangeId) {
-        // TODO : 에러처리
-        Exchange exchange = exchangeRepository.findExchangeById(exchangeId).orElseThrow();
-        return ExchangeDetailResponse.from(exchange);
+    private Exchange getById(Long exchangeId) {
+        return exchangeRepository.findExchangeById(exchangeId)
+                .orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_EXCHANGE));
     }
 
+    private Member getMemberById(Long memberId) {
+        return memberRepository.findById(memberId)
+                .orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_MEMBER));
+    }
+
+    private Region getRegionById(Long regionId) {
+        return regionRepository.findById(regionId)
+                .orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_REGION));
+    }
 }
